修复一些BUG

+++ src/main/java/com/user/controller/ArticleController.java
@@ -1,12 +1,21 @@
 package com.user.controller;
 
 import com.dubbo.ElasticsearchService;
+import com.dubbo.commons.Const;
 import com.dubbo.commons.ImgResultDto;
 import com.dubbo.commons.ServerResponse;
 import com.dubbo.entity.Article;
+import com.user.entity.User;
 import com.user.service.ArticleService;
+import com.user.utils.JedisUtil;
+import com.user.utils.JsonUtil;
+import com.user.utils.JwtTokenUtil;
+import com.user.vo.GetElasticsearchResultVo;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
 import org.apache.dubbo.config.annotation.Reference;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -27,40 +36,77 @@ public class ArticleController {
     @CrossOrigin(origins = "*", allowCredentials = "true", allowedHeaders = "*", methods = RequestMethod.OPTIONS, maxAge = 3600)
     @RequestMapping(value = "uploadImage")
     public ImgResultDto uploadImage(@RequestParam("uploadImage") List<MultipartFile> list, HttpServletRequest request, HttpServletResponse response){
-        //这里是我在web中定义的两个初始化属性，保存目录的绝对路径和相对路径，你们可以自定义
-        String imgUploadAbsolutePath = request.getServletContext().getInitParameter("imgUploadAbsolutePath");
-        String imgUploadRelativePath = request.getServletContext().getInitParameter("imgUploadRelativePath");
-
         //允许任何请求访问
         response.setHeader("Access-Control-Allow-Origin", "http://localhost:8082");
 //        response.setHeader("Access-Control-Allow-Headers", "Content-Length,X-Requested-With,content-type");
         response.setHeader("Access-Control-Allow-Methods", "PUT,POST,GET,DELETE,OPTIONS");
 
         //服务曾处理数据，返回Dto
-        return articleService.uploadImage(list,imgUploadAbsolutePath,imgUploadRelativePath,1);
+        return articleService.uploadImage(list,1);
     }
 
     //发表文章
     @RequestMapping(value = "uploadArticle",method = RequestMethod.POST)
-    public ServerResponse uploadArticle(Article article){
+    public ServerResponse uploadArticle(HttpServletRequest httpServletRequest,String userName,String articleTitle,String articleContent){
+        if (!checkRole(httpServletRequest, userName)){
+            return ServerResponse.createByErrorMessage("你使用的用户名和jwt token不一致");
+        }
+        //使用userName从redis反序列化出user，在得到userId->article_user_id
+        Article article=new Article(
+                JedisUtil.getArticleId(),
+                JsonUtil.stringToObj(JedisUtil.getValue(userName), User.class).getUserId(),
+                articleTitle,
+                articleContent,
+                null,
+                null
+        );
+        //将articleId-articleUserId存入redis中  commentId-1 10
+        JedisUtil.setKey(Const.Role.COMMENT_ID +article.getArticleId(),article.getArticleUserId());
+
         return elasticsearchService.addArticle(article);
     }
 
     //删除文章
     @RequestMapping(value = "deleteArticle",method = RequestMethod.POST)
-    public ServerResponse deleteArticle(String articleId){
+    public ServerResponse deleteArticle(HttpServletRequest httpServletRequest,String userName,String articleId){
+        if (!checkRole(httpServletRequest, userName)){
+            return ServerResponse.createByErrorMessage("你使用的用户名和jwt token不一致");
+        }
+        //根据文章ID查出此文章所属用户ID，比较ID是否相同
+        String articleUserId=JedisUtil.getValue(Const.Role.COMMENT_ID+articleId);
+        if (!JsonUtil.stringToObj(JedisUtil.getValue(userName), User.class).getUserId().equals(articleUserId)){
+            return ServerResponse.createByErrorMessage("不要删除别人的文章");
+        }
         return elasticsearchService.deleteArticle(articleId);
     }
 
     //更新文章
     @RequestMapping(value = "updateArticle",method = RequestMethod.POST)
-    public ServerResponse updateArticle(Article article){
+    public ServerResponse updateArticle(HttpServletRequest httpServletRequest,String userName,Article article){
+        if (!checkRole(httpServletRequest, userName)){
+            return ServerResponse.createByErrorMessage("你使用的用户名和jwt token不一致");
+        }
+        //根据文章ID查出此文章所属用户ID，比较ID是否相同
+        String articleUserId=JedisUtil.getValue(Const.Role.COMMENT_ID+article.getArticleId());
+        if (!JsonUtil.stringToObj(JedisUtil.getValue(userName), User.class).getUserId().equals(articleUserId)){
+            return ServerResponse.createByErrorMessage("不要更新别人的文章");
+        }
+        article.setArticleUserId(JsonUtil.stringToObj(JedisUtil.getValue(userName), User.class).getUserId());
         return elasticsearchService.updateArticle(article);
     }
 
     //GET文章
     @RequestMapping(value = "getArticle",method = RequestMethod.POST)
-    public ServerResponse getArticle(String articleId){
+    public ServerResponse getArticle(HttpServletRequest httpServletRequest,String userName,String articleId){
+        if (!checkRole(httpServletRequest, userName)){
+            return ServerResponse.createByErrorMessage("你使用的用户名和jwt token不一致");
+        }
+        //根据文章ID查出此文章所属用户ID，比较ID是否相同
+        String articleUserId=JedisUtil.getValue(Const.Role.COMMENT_ID+articleId);
+        if (!JsonUtil.stringToObj(JedisUtil.getValue(userName), User.class).getUserId().equals(articleUserId)){
+            return ServerResponse.createByErrorMessage("通过这种方式只能得到自己的文章，不能得到别人的文章");
+        }
+
         return elasticsearchService.selectArticle(articleId);
     }
 
@@ -70,8 +116,30 @@ public class ArticleController {
      * 前端传过来    搜索人id,搜索关键字
      */
     @RequestMapping(value = "searchArticle",method = RequestMethod.POST)
-    public ServerResponse searchArticle(String userInputText,int pageNum){
+    public ServerResponse searchArticle(HttpServletRequest httpServletRequest,String userName,String userInputText,int pageNum){
+        if (!checkRole(httpServletRequest, userName)){
+            return ServerResponse.createByErrorMessage("你使用的用户名和jwt token不一致");
+        }
         return elasticsearchService.searchArticle(userInputText,pageNum);
     }
 
+
+
+
+
+
+    private Boolean checkRole(HttpServletRequest httpServletRequest,String userName){
+        String token = httpServletRequest.getHeader(JwtTokenUtil.tokenHeader);
+        if (!StringUtils.isEmpty(token)) {
+            Claims claims = Jwts.parser()
+                    .setSigningKey(JwtTokenUtil.secret)
+                    .parseClaimsJws(token.replace(JwtTokenUtil.tokenPrefix, ""))
+                    .getBody();
+            //相等jwt中token和用户名一致
+            if (claims.getSubject().equals(userName)){
+                return true;
+            }
+        }
+        return false;
+    }
 }
